+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Additions and Improvements
 - Execution layer (The Merge):
   - Execution specific RPC endpoint [[#3378](https://github.com/hyperledger/besu/issues/3378)
+  - Adds JWT authentication to Engine APIs
 - Tracing APIs: trace_rawTransaction, trace_get, trace_callMany
 
 ### Bug Fixes